+++ datops-frontend/src/main/live_opts/core.cljs
@@ -71,6 +71,10 @@
    [:regularmarketprice "Stock Market Price" "SMP"]
    [:regularmarketdaylow "Stock Market Day Low" "SMDL"]
    [:regularmarketdayhigh "Stock Market Day High" "SMDH"]
+   [:delta "Delta" "δ"]
+   [:gamma "Gamma" "γ"]
+   [:theta "Theta" "θ"]
+   [:vega "Vega" "ν"]
    ;;[:quotetype "Quote Type" "QT"]
    [:lastcrawl "Last Updated" "LU"]
    ])
@@ -114,10 +118,6 @@
         min-sto (.-value (gdom/getElement "min-sto-value"))
         max-sto (.-value (gdom/getElement "max-sto-value"))
 
-        ;; Premium/strike ratio
-        min-pso (.-value (gdom/getElement "min-pso-value"))
-        max-pso (.-value (gdom/getElement "max-pso-value"))
-
         ;; Yield
         min-yield (.-value (gdom/getElement "min-yield-value"))
         max-yield (.-value (gdom/getElement "max-yield-value"))
@@ -126,6 +126,16 @@
         min-myield (.-value (gdom/getElement "min-myield-value"))
         max-myield (.-value (gdom/getElement "max-myield-value"))        
 
+        ;; Greeks
+        min-delta (.-value (gdom/getElement "min-delta-value"))
+        max-delta (.-value (gdom/getElement "max-delta-value"))
+        min-gamma (.-value (gdom/getElement "min-gamma-value"))
+        max-gamma (.-value (gdom/getElement "max-gamma-value"))
+        min-theta (.-value (gdom/getElement "min-theta-value"))
+        max-theta (.-value (gdom/getElement "max-theta-value"))
+        min-vega (.-value (gdom/getElement "min-vega-value"))
+        max-vega (.-value (gdom/getElement "max-vega-value"))
+        
         ;; Market capitalization
         min-cap (.-value (gdom/getElement "min-cap-value"))
         max-cap (.-value (gdom/getElement "max-cap-value"))
@@ -173,10 +183,6 @@
           "&min-sto=" (js/encodeURIComponent min-sto)
           "&max-sto=" (js/encodeURIComponent max-sto)
 
-          ;; Premium/strike ratio
-          "&min-pso=" (js/encodeURIComponent min-pso)
-          "&max-pso=" (js/encodeURIComponent max-pso)
-
           ;; Yield
           "&min-yield=" (js/encodeURIComponent min-yield)
           "&max-yield=" (js/encodeURIComponent max-yield)
@@ -184,6 +190,16 @@
           ;; Monthly yield
           "&min-myield=" (js/encodeURIComponent min-myield)
           "&max-myield=" (js/encodeURIComponent max-myield)
+
+          ;; Greeks
+          "&min-delta=" (js/encodeURIComponent min-delta)
+          "&max-delta=" (js/encodeURIComponent max-delta)
+          "&min-gamma=" (js/encodeURIComponent min-gamma)
+          "&max-gamma=" (js/encodeURIComponent max-gamma)
+          "&min-theta=" (js/encodeURIComponent min-theta)
+          "&max-theta=" (js/encodeURIComponent max-theta)
+          "&min-vega=" (js/encodeURIComponent min-vega)
+          "&max-vega=" (js/encodeURIComponent max-vega)
 
           ;; Market cap
           "&min-cap=" (js/encodeURIComponent min-cap)
@@ -209,13 +225,19 @@
                      :puts puts :calls calls
                      :stock stock :etf etf
                      :min-sto min-sto :max-sto max-sto
-                     :min-pso min-pso :max-pso max-pso
                      :min-yield min-yield :max-yield max-yield
                      :min-myield min-myield :max-myield max-myield
+                     :min-delta min-delta :max-delta max-delta
+                     :min-gamma min-gamma :max-gamma max-gamma
+                     :min-theta min-theta :max-theta max-theta
+                     :min-vega min-vega :max-vega max-vega
                      :min-cap min-cap :max-cap max-cap
                      :order-by order-by :limit limit :active active
                      })}}))
-            {:keys [quotes options catalysts]} (-> resp :body read-string)]
+            {:keys [quotes options catalysts]} (-> resp :body read-string)
+            quotes (if (contains? quotes :error)
+                     []
+                     quotes)]
         (swap! state #(-> %
                           (assoc :cur-quotes quotes)
                           (assoc :cur-results options)
@@ -375,7 +397,7 @@
          (keep-indexed
           (fn [idx [id c-name short-name]]
             [:th
-             {:key id :class [(name id)] :title c-name}
+             {:key id :class [(name id)] :title c-name :scope "col"}
              [:p
               {:on-click
                (fn [_]
@@ -403,7 +425,7 @@
                (doall
                 (map
                  (fn [[id _]]
-                   (let [v (get entry id)]
+                   (let [v (get entry id nil)]
                      [:td {:key (str contractsymbol "-" (name id))
                            :class [(name id)
                                    (cond
@@ -415,10 +437,17 @@
 
                              (or (= id :impliedvolatility)
                                  (= id :yield) (= id :monthlyyield))
-                             (gstring/format "%.2f" v)
+                             (if (number? v)
+                               (gstring/format "%.2f" v)
+                               (str v))
+
+                             (contains? #{:delta :theta :gamma :vega} id)
+                             (if (number? v) (gstring/format "%.4f" v) "")
                              
                              (or (= id :expiration) (= id :lasttradedate))
-                             (-> (from-ts (+ (or v 0) offset-exp)) (str/split #",") first)
+                             (if (number? v)
+                               (-> (from-ts (+ (or v 0) offset-exp)) (str/split #",") first)
+                               (str v))
 
                              (or (= id :ask) (= id :bid) (= id :lastprice)
                                  (= id :regularmarketprice)
@@ -468,8 +497,14 @@
     (.addEventListener
      el "keydown"
      (fn [ev] (when (= (.-keyCode ev) 13) (send-query state)))))
-  (doseq [el (.getElementsByTagName js/document "button")]
-    (.addEventListener el "click" (fn [] (send-query state))))
+  (.addEventListener (gdom/getElement "send") "click" (fn [] (send-query state)))
+  (.addEventListener
+   (gdom/getElement "clear") "click"
+   (fn []
+     (doseq [el (.getElementsByTagName js/document "input")]
+       (if (= "checkbox" (.-type el))
+         (set! (.-checked el) "")
+         (set! (.-value el) "")))))
   (let [q-string (.. js/window -location -search)
         url-params (js->clj (new js/URLSearchParams q-string))
         order-by (.get url-params "order-by")
@@ -558,14 +593,6 @@
       (when max-sto
         (set! (.-value (gdom/getElement "max-sto-value")) max-sto)))
 
-    ;; Premium/Strike
-    (let [min-pso (.get url-params "min-pso")
-          max-pso (.get url-params "max-pso")]
-      (when min-pso
-        (set! (.-value (gdom/getElement "min-pso-value")) min-pso))
-      (when max-pso
-        (set! (.-value (gdom/getElement "max-pso-value")) max-pso)))
-
     ;; Yield
     (let [min-yield (.get url-params "min-yield")
           max-yield (.get url-params "max-yield")]
@@ -581,6 +608,32 @@
         (set! (.-value (gdom/getElement "min-myield-value")) min-myield))
       (when max-myield
         (set! (.-value (gdom/getElement "max-myield-value")) max-myield)))
+
+    ;; greeks
+    (let [min-delta (.get url-params "min-delta")
+          max-delta (.get url-params "max-delta")
+          min-gamma (.get url-params "min-gamma")
+          max-gamma (.get url-params "max-gamma")
+          min-theta (.get url-params "min-theta")
+          max-theta (.get url-params "max-theta")
+          min-vega (.get url-params "min-vega")
+          max-vega (.get url-params "max-vega")]
+      (when min-delta
+        (set! (.-value (gdom/getElement "min-delta-value")) min-delta))
+      (when max-delta
+        (set! (.-value (gdom/getElement "max-delta-value")) max-delta))
+      (when min-gamma
+        (set! (.-value (gdom/getElement "min-gamma-value")) min-gamma))
+      (when max-gamma
+        (set! (.-value (gdom/getElement "max-gamma-value")) max-gamma))
+      (when min-theta
+        (set! (.-value (gdom/getElement "min-theta-value")) min-theta))
+      (when max-theta
+        (set! (.-value (gdom/getElement "max-theta-value")) max-theta))
+      (when min-vega
+        (set! (.-value (gdom/getElement "min-vega-value")) min-vega))
+      (when max-vega
+        (set! (.-value (gdom/getElement "max-vega-value")) max-vega)))
 
     ;; Market cap
     (let [min-cap (.get url-params "min-cap")